Add files via upload
We all must have seen a drum kit in some concert or elsewhere it is a collection of drums, cymbals and other percussion instruments. Well, so here we are with the goal for this project to build a drum kit app that runs in the browser. The main concepts presented in the article are DOM, key events and CSS animations.


So the way this website would work that we’ve got a number of keys on the webpage that represents different drums in a typical drum set and when you click on any of those buttons then you’ll get the corresponding sound of the drum. And in addition you can also use the keys on the keyboard to have a sound effect. we can also change the theme of the website as well as the background of the website.


The languages we are going to use in the projects are as followed:-

HTML(HYPER TEXT MARKUP LANGUAGE)
CSS(CASCADING STYLE SHEET)
JAVASCRIPT